requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt

# WSGI Handler
# ------------------------------------------------
gevent==1.1.2
gunicorn==19.6.0

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==0.4.2

# Static and Media Storage
# ------------------------------------------------
boto3==1.4.7
django-storages-redux==1.3.2
django-storages==1.6.5
Collectfast==0.2.3
